test: split admin workflow tests
